Set correct error codes for Hanami app exceptions (#7)
The BetterErrors middleware always returns a 500 status when rescuing an exception (outside of Rails). This is not not always appropriate, such as for a `Hanami::Router::NotFoundError`, which should be a 404.

To account for this, gently patch `BetterErrors::Middleware#show_error_page` (which is called only when an exception has been rescued) to pass that rescued exception to a proc we inject into the rack env here in our own middleware. This allows our middleware to know the about exception class and provide the correct status code after BetterErrors is done with its job.

This is a monkey-patch, yes, but BetterErrors is very mature gem right now, with most of its code unchanged in 10 years. I'm comfortable with this arrangement until we can either (a) introduce some change to BetterErrors itself to let us directly control the status code, or (b) introduce our own Hanami-owned replacement to BetterErrors. Both of these would be well after Hanami v2.2.